refactor: move all interactive keyboard code to dedicated `interactive` folder (Android)
