+++ android/src/main/java/com/reactnativekeyboardcontroller/interactive/InteractiveKeyboardProvider.kt
@@ -1,4 +1,4 @@
-package com.reactnativekeyboardcontroller
+package com.reactnativekeyboardcontroller.interactive
 
 object InteractiveKeyboardProvider {
   var shown = false
