+++ android/src/main/java/com/reactnativekeyboardcontroller/interactive/KeyboardAnimationController.kt
@@ -1,4 +1,4 @@
-package com.reactnativekeyboardcontroller
+package com.reactnativekeyboardcontroller.interactive
 
 import android.os.CancellationSignal
 import android.view.View
